Fix/config sanity (#21)
* fix(config): task-scope resolved logging and contrastive pretraining prob

* fix(logging): print resolved config and clarify MLM corruption mix

* refactor(logging): replace %-style formatting with f-strings

* fix(config): standardize CLI bool parsing

* fix(config): add semantic validation and explicit wandb enable

* fix(config): deprecate unused scheduler.num_cycles

* fix(glue): keep canonical task and task-scoped wandb config

* fix(pretraining): require explicit streaming eval budget

* fix(logging): trim dead keys from task-scoped config payloads

* docs(config): document explicit eval budget and legacy key deprecations

* fix(tokenizer): require explicit special-token rewrite opt-in

* docs(config): remove misleading pretraining use_deepspeed example

* chore(config): set pretraining mlm_probability to 0.25

* refactor(logging): remove remaining logger percent placeholders

* fix(config): harden schema validation and task-scoped payloads

* fix(trainer): wire cpu and tokenization config knobs

* chore(logging): replace printf-style compile logs

* fix(logging): drop pretraining no-op trainer keys from payload

* fix(config): allow zero checkpoint retention limits

* chore(configs): remove deprecated report_to in pretraining presets

* fix(pretraining): streamline train metrics and restore accuracy defaults

* feat(logging): default wandb watch to gradients for online runs

* chore(configs): surface key defaults in pretraining and glue presets

* fix(training): make accelerator init resilient across task trainers

* fix(contrastive): harden batch flow and iterator lifecycle

* fix(pretraining): fail fast on empty dataloader epochs

* fix(pretraining): unify checkpoint retention and save layout

* chore(checkpoints): switch consumers to unified checkpoints tree

* fix(config): centralize mixed precision and remove GLUE task mutation

* docs: consolidate canonical docs and fix stale training/config guidance

* docs(training): add practical pretraining crash-recovery resume guide

* feat(pretraining): enable best-effort streaming resume

Also enforces bf16/no mixed-precision policy (fp16 unsupported) and updates docs/tests.

* feat(logging): compact terminal-aware resolved config display

* fix(training): honor tqdm flag and wire mteb task overrides

Also makes save/eval step validation strategy-aware with tests and doc updates.

* fix: restore portable distributed checkpoint compatibility

* feat(config): add yaml variables and dot-path load overrides

* docs(dev): add deferred strict config-loader architecture todos

* docs(config): add shared seq_len variables example

* docs: format


* fix: harden distributed checkpoint export and mteb loading

* chore(config): align checkpoint paths and streaming eval budget

* fix(glue): fallback to deepspeed shards when safetensors missing

* perf(config): disable train accuracy in 100m muonclip preset

* perf(config): disable train accuracy logging by default

* feat(logging): align wrapped resolved-config lines

* fix(contrastive): guard checkpoint retention against null limits

* fix(glue): allow hub tokenizers without mlm mask tokens

* fix(precision): enforce fp16 policy across runtime and export

* chore(wip): checkpoint in-progress review remediation

* fix(glue): honor keep-all retention and hf input semantics

* fix(contrastive): use canonical checkpoint config fields

* test(pretraining): cover epoch-local streaming resume skip

* fix(glue): pass token_type_ids in training forward

* fix(config): allow save_steps=0 when save strategy is no

* fix(training): enforce MuonClip distributed compatibility

* docs(model): clarify RoPE fp32 and Muon full-tensor assumptions

* fix(checkpointing): unify checkpoint root handling for glue and contrastive

* fix(training): harden checkpoint pruning and clarify review false-positives

* fix(glue): isolate train/eval metric state and clarify invariants

* fix(contrastive): correct checkpoint retention off-by-one

* fix(checkpointing): align contrastive retention pruning

* fix(config): preserve larger tokenizer max length for staged training

* docs(checkpointing): clarify latest resume semantics

* fix(glue): align distributed schedule, resume state, and sweep config

* fix(mteb): load safetensors checkpoints with non-strict matching

* fix(config): accept wandb.watch weights alias

* fix(mteb): fail fast on checkpoint mismatch and empty tasks

* fix(glue): unify checkpoint writes and harden distributed eval IO

* fix(glue,config): harden cli bool overrides and glue runtime validation

* fix(mteb,preprocess): honor tokenizer overrides and robust latest step

* 🎨


---------

+++ src/neobert/pretraining/masked_objective.py
@@ -136,7 +136,9 @@ def streaming_argmax(
         for vocab_start in range(0, vocab_size, vocab_chunk):
             vocab_end = min(vocab_start + vocab_chunk, vocab_size)
             weight_chunk = weight[vocab_start:vocab_end].to(accum_dtype)
-            logits_chunk = hidden_chunk @ weight_chunk.t()
+            # Under outer autocast contexts, matmul can still return lower-precision
+            # tensors; cast explicitly so indexed updates stay dtype-consistent.
+            logits_chunk = (hidden_chunk @ weight_chunk.t()).to(accum_dtype)
             chunk_values, chunk_argmax = logits_chunk.max(dim=1)
             better = chunk_values > best_values
             if better.any():
@@ -207,7 +209,9 @@ def streaming_ce_sum(
         for vocab_start in range(0, vocab_size, vocab_chunk):
             vocab_end = min(vocab_start + vocab_chunk, vocab_size)
             weight_chunk = weight[vocab_start:vocab_end].to(accum_dtype)
-            logits_chunk = hidden_chunk @ weight_chunk.t()
+            # Keep streaming CE numerics in ``accum_dtype`` even when called under
+            # autocast-enabled training contexts.
+            logits_chunk = (hidden_chunk @ weight_chunk.t()).to(accum_dtype)
 
             in_chunk = (target_chunk >= vocab_start) & (target_chunk < vocab_end)
             if in_chunk.any():
